Make in-call speech recognition failures non-fatal
会話中に SFSpeechRecognizer が一時的に失敗（無音時の No speech detected 等）
すると即「エラーなのだ」で通話が終了していた。最大3回聞き直し、それでも
聞き取れない場合はエラーにせず挨拶して通話を終える。認識まわりは
CallViewModel+SpeechRecognition.swift に分離（file_length 対応）。
会話高速化（無音判定短縮・短文化）は #103 に分離し、本PRには含めない。
